Cleanup Base pallet Config and Event types.
